+++ internal/vba/cfg/clone.go
@@ -54,6 +54,7 @@ func cloneParameters(in []procedureir.Parameter) []procedureir.Parameter {
 	for i := range out {
 		out[i].PassingRange = cloneRangePointer(in[i].PassingRange)
 		out[i].DefaultRange = cloneRangePointer(in[i].DefaultRange)
+		out[i].NameRange = cloneRangePointer(in[i].NameRange)
 		out[i].BoundsRange = cloneRangePointer(in[i].BoundsRange)
 		out[i].ArrayBounds = cloneArrayBounds(in[i].ArrayBounds)
 	}