+++ internal/vba/procedureir/clone.go
@@ -96,6 +96,7 @@ func cloneParameters(in []Parameter) []Parameter {
 	for i := range out {
 		out[i].PassingRange = cloneRangePointer(in[i].PassingRange)
 		out[i].DefaultRange = cloneRangePointer(in[i].DefaultRange)
+		out[i].NameRange = cloneRangePointer(in[i].NameRange)
 		out[i].BoundsRange = cloneRangePointer(in[i].BoundsRange)
 		out[i].ArrayBounds = append([]ArrayBound(nil), in[i].ArrayBounds...)
 		for j := range out[i].ArrayBounds {